feat: removed custom request context creationa and moved the getFilesInAdminUiFolder method to util
