+++ packages/vendure-plugin-modify-customer-orders/test/modify-customer-orders.spec.ts
@@ -16,7 +16,7 @@ import { expect, describe, beforeAll, afterAll, it, vi, test } from 'vitest';
 import path from 'path';
 import * as fs from 'fs';
 import { compileUiExtensions } from '@vendure/ui-devkit/compiler';
-import getFilesInAdminUiFolder from '../../util/src/compile-admin-ui.util';
+import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
 
 describe('Customer managed groups', function () {
   let server: TestServer;
