+++ packages/vendure-plugin-stripe-subscription/test/stripe-subscription.spec.ts
@@ -60,7 +60,7 @@ import {
 } from './helpers';
 import { expect, describe, beforeAll, afterAll, it, vi, test } from 'vitest';
 import { gql } from 'graphql-tag';
-import getFilesInAdminUiFolder from '../../util/src/compile-admin-ui.util';
+import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
 
 describe('Stripe Subscription Plugin', function () {
   let server: TestServer;
