+++ packages/vendure-plugin-webhook/test/e2e.spec.ts
@@ -16,6 +16,7 @@ import {
   SqljsInitializer,
   testConfig,
 } from '@vendure/testing';
+import { getSuperadminContext } from '@vendure/testing/lib/utils/get-superadmin-context';
 import { TestServer } from '@vendure/testing/lib/test-server';
 import nock from 'nock';
 import { initialData } from '../../test/src/initial-data';
@@ -35,7 +36,7 @@ import {
 } from '../src/ui/queries';
 import { stringifyProductTransformer } from './test-helpers';
 import { describe, beforeAll, it, expect, afterEach, afterAll } from 'vitest';
-import getFilesInAdminUiFolder from '../../util/src/compile-admin-ui.util';
+import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
 
 describe('Webhook plugin', function () {
   let server: TestServer;
@@ -78,12 +79,7 @@ describe('Webhook plugin', function () {
       productsCsvPath: '../test/src/products-import.csv',
     });
     serverStarted = true;
-    ctx = new RequestContext({
-      apiType: 'admin',
-      channel: await server.app.get(ChannelService).getDefaultChannel(),
-      isAuthorized: true,
-      authorizedAsOwnerOnly: false,
-    });
+    ctx = await getSuperadminContext(server.app);
   }, 60000);
 
   // Clear nock mocks after each test
